Database remodelling test

+++ Project Files/website/views.py
@@ -2,8 +2,8 @@ from flask import Blueprint, render_template, request, flash
 from flask_login import login_required, current_user
 import feedparser
 from . import db
-from .db_models import RSS_Data, Readlist, Tags
-from .rec_alg import recommendation_algorithm, tag_counter, global_tag_counter
+from .db_models import RSS_Data, Readlist, Tags, User_Website_Link, User_Readlist_Link
+from .external_functions import top_interaction_recommendation_algorithm, tag_counter, global_tag_counter
 views = Blueprint('views', __name__)
 
 
@@ -21,62 +21,64 @@ def home():
     global_tags = Tags.query.all()
 
     user_saved_websites = current_user.rss_data
-    print(user_saved_websites)
 
-    # for user_web in user_saved_websites:
-    #     web_clicks = {'web_id': user_web.id,
-    #                   'click_count':user_web.clicks}
-    #     click_counts.append(web_clicks)
-    #     tags_format = {'tag_id': user_web.tag_id,
-    #                    'tag_name': f'{global_tags[user_web.tag_id - 1].name}'}
-    #     if tags_format not in user_tags:
-    #         user_tags.append(tags_format)
-    #
+    for user_web in user_saved_websites:
+        web_clicks = {'web_id': user_web.rss_data.id,
+                      'click_count': user_web.clicks}
+        click_counts.append(web_clicks)
+        tags_format = {'tag_id': user_web.rss_data.tag_id,
+                       'tag_name': f'{global_tags[user_web.rss_data.tag_id].name}'}
+        if tags_format not in user_tags:
+            user_tags.append(tags_format)
 
-    # if request.method == 'POST':
-    #     website_id = request.form.get('website_id')
-    #     print(website_id)
-    #     live_click_dictionary = None
-    #     for click_dictionary in click_counts:
-    #         if click_dictionary['web_id'] == int(website_id):
-    #             live_click_dictionary = click_dictionary
-    #
-    #     website = RSS_Data.query.filter_by(id=website_id, user_id=current_user.id).first()
-    #     website.clicks = int(live_click_dictionary['click_count']) + 1
-    #
-    #     db.session.commit()
-    #     if "feed_id" in request.form:
-    #
-    #
-    #         website_title = request.form.get('feed_id')
-    #         website = RSS_Data.query.filter_by(title=website_title, user_id=current_user.id).first()
-    #
-    #         if website:
-    #             website_link = website.link
-    #         else:
-    #             print("No websites stored in database.")
-    #             website_link = "None"
-    #
-    #     if "art_tag_id" in request.form:
-    #         tag_id = request.form.get('art_tag_id')
-    #         title = request.form.get('article_title')
-    #         description = request.form.get('article_desc')
-    #         article_link = request.form.get('article_link')
-    #         print(tag_id)
-    #
-    #
-    #         existing_article = Readlist.query.filter_by(art_title=title).first()
-    #         if existing_article and existing_article.user_id == current_user.id:
-    #             flash('Article already saved', category='error')
-    #         else:
-    #
-    #
-    #
-    #
-    #             saved_article = Readlist(art_title=title, art_desc=description, art_link=article_link,tag=tag_id, user_id=current_user.id)
-    #             print(saved_article)
-    #             db.session.add(saved_article)
-    #             db.session.commit()
+
+    if request.method == 'POST':
+        website_id = request.form.get('website_id')
+
+        live_click_dictionary = None
+        for click_dictionary in click_counts:
+            if click_dictionary['web_id'] == int(website_id):
+                live_click_dictionary = click_dictionary
+        website = User_Website_Link.query.filter_by(rss_data_id=website_id, user_id=current_user.id).first()
+        website.clicks = int(live_click_dictionary['click_count']) + 1
+        db.session.commit()
+
+        if "feed_link" in request.form:
+            website_link = request.form.get('feed_link')
+            website = RSS_Data.query.filter_by(link=website_link).first()
+
+            if website:
+                website_link = website.link
+            else:
+                print("No websites stored in database.")
+                website_link = "None"
+
+        if "art_tag_id" in request.form:
+            tag_id = request.form.get('art_tag_id')
+            title = request.form.get('article_title')
+            description = request.form.get('article_desc')
+            article_link = request.form.get('article_link')
+
+
+            existing_article = Readlist.query.filter_by(art_title=title).first()
+
+            if not existing_article:
+                tag_instance = Tags.query.get(tag_id)
+                saved_article = Readlist(art_title=title, art_desc=description, art_link=article_link, tag=tag_instance)
+                db.session.add(saved_article)
+                db.session.commit()
+                existing_article = saved_article
+            user_existing_article = False
+            for article in current_user.readlist:
+                if article.readlist.art_title == title:
+                    user_existing_article = True
+            if user_existing_article == False:
+                new_user_entry = User_Readlist_Link(user_id=current_user.id, readlist_id=existing_article.id)
+                db.session.add(new_user_entry)
+                db.session.commit()
+            elif user_existing_article == True:
+                flash('Article already saved', category='error')
+
 
     feed = feedparser.parse(website_link)
 
@@ -90,10 +92,16 @@ def delete_website():
     if request.method == 'POST':
         if "delete_link" in request.form:
             website_id = request.form.get('id')
-            website = RSS_Data.query.get(website_id)
-            if website:
-                if website.user_id == current_user.id:
-                    db.session.delete(website)
+            link_table_website = User_Website_Link.query.filter_by(rss_data_id=website_id, user_id=current_user.id).first()
+            print(link_table_website)
+            if link_table_website:
+                if link_table_website.user_id == current_user.id:
+                    website = RSS_Data.query.get(website_id)
+                    if len(website.users) == 1:
+                        db.session.delete(website)
+
+                    db.session.delete(link_table_website)
+
                     db.session.commit()
                     flash('Website deleted successfully', category='success')
             else:
@@ -108,12 +116,12 @@ def add_links():
     user_tags = []
 
     tags_list = Tags.query.all()
-    user_saved_websites = RSS_Data.query.filter_by(user_id=current_user.id).all()
+    user_saved_websites = current_user.rss_data
 
     for user_web in user_saved_websites:
 
-        tags_format = {'tag_id': user_web.tag_id,
-                       'tag_name': f'{tags_list[user_web.tag_id - 1].name}'}
+        tags_format = {'tag_id': user_web.rss_data.tag_id,
+                       'tag_name': f'{tags_list[user_web.rss_data.tag_id].name}'}
         if tags_format not in user_tags:
             user_tags.append(tags_format)
 
@@ -124,15 +132,31 @@ def add_links():
             website_link = request.form.get('web_link')
             website_tag = request.form.get('web_tag')
             print(website_tag)
-            website_existence = RSS_Data.query.filter_by(title=website_title, user_id=current_user.id).first()
-            if website_existence and current_user.id == website_existence.user_id:
-                flash('This title already exists', category='error')
-            else:
-                new_link = RSS_Data(title=website_title, link=website_link, clicks=0, tag_id=website_tag, user_id=current_user.id)
-                db.session.add(new_link)
-                db.session.commit()
 
-    return render_template("website_add.html", user=current_user,tags=tags_list, user_tags=user_tags)
+            website_existence = RSS_Data.query.filter_by(title=website_title).first()
+
+            try:
+                website_link_existence = User_Website_Link.query.filter_by(user_id=current_user.id, rss_data_id=website_existence.id).first()
+            except AttributeError:
+                website_link_existence = False
+
+
+            if website_existence and website_link_existence:
+                flash('This title already exist | It\'s already linked to your account', category='error')
+            elif website_existence and website_link_existence == None:
+                new_website_link = User_Website_Link(user_id=current_user.id, rss_data_id=website_existence.id, clicks=0)
+                db.session.add(new_website_link)
+                flash('This title has been added', category='success')
+            else:
+                new_website = RSS_Data(title=website_title, link=website_link, tag_id=website_tag)
+                db.session.add(new_website)
+                db.session.commit()
+                new_website_link = User_Website_Link(user_id=current_user.id, rss_data_id=new_website.id, clicks=0)
+                db.session.add(new_website_link)
+            db.session.commit()
+
+
+    return render_template("website_add.html", user=current_user, tags=tags_list, user_tags=user_tags)
 
 
 @views.route('/read_later', methods=['GET', 'POST'])
@@ -143,30 +167,37 @@ def read_later():
     # Gets all user tags related to the articles that the user has saved so that only those tags are presented to the user.
     user_tags = []
     global_tags = Tags.query.all()
-    user_saved_articles = Readlist.query.filter_by(user_id=current_user.id).all()
-
+    user_saved_articles = current_user.readlist
     for user_art in user_saved_articles:
-        print(type(user_art.tag))
-        tags_format = {'tag_id': user_art.tag,
-                       'tag_name': f'{global_tags[(user_art.tag) - 1].name}'}
+        tags_format = {'tag_id': user_art.readlist.tag_id,
+                       'tag_name': f'{global_tags[(user_art.readlist.tag_id)].name}'}
         if tags_format not in user_tags:
             user_tags.append(tags_format)
-    # --- POST METHOD --- #
+    #--- POST METHOD --- #
     if request.method == 'POST':
         # --- Removes article that is saved from readlist --- #
         if "unlike_article" in request.form:
             article_id = request.form.get('unlike_article')
             article_to_del = Readlist.query.get(article_id)
-            db.session.delete(article_to_del)
+
+            readlist_linktable_entry = User_Readlist_Link.query.filter_by(user_id=current_user.id, readlist_id=article_to_del.id).first()
+            print(article_to_del.users)
+            db.session.delete(readlist_linktable_entry)
+
+            if len(article_to_del.users) == 0: # 0 value since the linked user is deleted beforehand, so it will have an empty associated list.
+                db.session.delete(article_to_del)
             db.session.commit()
+
         # --- Refreshes the webpage with the relevant filters --- #
         if "filter_article" in request.form:
             tag_filter = request.form.get('filter_article')
-            print(tag_filter)
-            user_tags = [{'tag_id': int(tag_filter),
-                          'tag_name': f'{global_tags[int(tag_filter)-1].name}'}]
-            print(user_tags)
+            if tag_filter != None:
+                tag_filter = int(tag_filter)
 
+            user_tags = [{'tag_id': tag_filter,
+                          'tag_name': f'{global_tags[tag_filter-1].name}'}]
+            print(user_tags)
+        ...
     return render_template("read_later.html", user=current_user, tags=user_tags)
 
 
@@ -174,7 +205,7 @@ def read_later():
 @login_required
 def discover():
     # Runs discovery algorithm to prioritise tags
-    user_recommended_tags = recommendation_algorithm()
+    user_recommended_tags = top_interaction_recommendation_algorithm()
     user_websites_sorted = []
 
     # --- Separation for all the POST method code --- #
@@ -182,25 +213,44 @@ def discover():
         if "article_title" in request.form:
             title = request.form.get('article_title')
             existing_article = Readlist.query.filter_by(art_title=title).first()
-            if existing_article and existing_article.user_id == current_user.id:
+            print(existing_article)
+            try:
+                readlist_link = User_Readlist_Link.query.filter_by(readlist_id=existing_article.id, user_id=current_user.id).first()
+            except AttributeError:
+                readlist_link = None
+            print(readlist_link)
+            if existing_article and readlist_link:
+                # If statement that evaluates if article and readlist exists
+                print("True")
                 flash('Article already saved', category='error')
+            elif existing_article and readlist_link == None:
+                print(True)
+                # Function that finds an existing article and adds it to the users articles.
+                new_readlist_link = User_Readlist_Link(user_id=current_user.id, readlist_id=existing_article.id)
+                db.session.add(new_readlist_link)
+                flash('Article saved', category='success')
             else:
 
                 link = request.form.get('article_link')
                 description = request.form.get('article_desc')
-                saved_article = Readlist(art_title=title, art_desc=description, art_link=link, user_id=current_user.id)
+                article_id = request.form.get('article_id')
+                saved_article = Readlist(art_title=title, art_desc=description, art_link=link, tag_id=article_id)
                 db.session.add(saved_article)
                 db.session.commit()
+                new_readlist_link = User_Readlist_Link(user_id=current_user.id, readlist_id=saved_article.id)
+                db.session.add(new_readlist_link)
+                flash('Article saved', category='success')
+            db.session.commit()
 
         if "add_discovery_feed" in request.form:
-            article_to_add = RSS_Data(title=request.form.get('add_discovery_feed'),
-                                      link=request.form.get('source_link'), clicks=0, tag_id=request.form.get('source_tag'), user_id=current_user.id)
-            db.session.add(article_to_add)
+            website_to_add = RSS_Data.query.filter_by(title=request.form.get('add_discovery_feed')).first()
+            website_link_to_add = User_Website_Link(user_id=current_user.id, rss_data_id=website_to_add.id, clicks=0)
+            db.session.add(website_link_to_add)
             db.session.commit()
         if "filter_websites" in request.form:
             user_desired_tag = request.form.get("filter_websites")
             if user_desired_tag == "None":
-                user_recommended_tags = recommendation_algorithm()
+                user_recommended_tags = top_interaction_recommendation_algorithm()
             else:
 
                 user_recommended_tags = [user_desired_tag]
@@ -213,14 +263,13 @@ def discover():
         individual_tag_groups = [[relevant_tag.id, relevant_tag.name]]
 
         for website in global_websites:
+            website_existing = False
             # Checks if user has already added the website to their personal feed. IF they have, it does not appear.
-            if website.user_id is None:
-                website_existing = False
-                for saved_website in current_user.rss_data:
-                    if saved_website.title == website.title:
-                        website_existing = True
+            for website_links in website.users:
+                if website_links.user_id == current_user.id:
+                    website_existing = True
 
-            if website_existing != True:
+            if not website_existing:
 
                 if website.tag_id == int(user_tag):
 
@@ -254,9 +303,7 @@ def user_stats():
     global_tags_id = [tag.id for tag in global_tags]
     stats_data = tag_counter()
     global_tag_clicks = global_tag_counter()
-    print(global_tag_clicks)
-    user_websites = RSS_Data.query.filter_by(user_id=current_user.id).order_by(RSS_Data.clicks.desc()).limit(4)
-
+    user_websites = User_Website_Link.query.filter_by(user_id=current_user.id).order_by(User_Website_Link.clicks.desc()).limit(4)
